==> 05_INT_by_task.Rmd <==
---
title: "INT_by_task"
author: "Shane Devine McKeon"
output: html_document

---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(LNCDR)
library(dplyr)
library(ggplot2)
library(mgcv)
library(lme4)
library(lmerTest)
library(tidyr)
library(gratia)
library(tidyverse)
library(ggeffects)
library(marginaleffects)
library(extrafont)
extrafont::loadfonts()
library(conflicted)
library(eegkit)

conflict_prefer("filter", "dplyr")  # Always use dplyr::filter()


```


# Regions 

## EEG Fig
```{r}
chanLocs <- read.csv('/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/ChannelLocs.csv')
F_labels <- chanLocs$labels[grep("^F", chanLocs$labels)]
A_labels <- chanLocs$labels[grep("^A", chanLocs$labels)]
P_labels <- chanLocs$labels[grep("^P", chanLocs$labels)]
O_labels <- chanLocs$labels[grep("^O", chanLocs$labels)]


pdf("/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure2/frontalEEGplot.pdf", width = 5, height = 5)

eegkit::eegcap(electrodes = c("Fp1", "F7", "F5", "F3", "F1", "FC1", "FC3", "FC5", "FT9", "Fp2", "F8", "F6", "F4", "F2", "Fz", "FCz", "FC2", "FC4", "FC6", "FT10", #Frontal
                              "AF1", "AF5", "AFz", "AF2", "AF6","Cz", "C1", "C2"),
                              col.point = c("#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d", "#012b0d","#012b0d","#012b0d"),
                             cex.point = 3, nose = TRUE, ears = TRUE,
               plotlabels = FALSE,  plotaxes = FALSE, col.head = "grey", type="2d")

dev.off()
```

```{r}

pdf("/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure2/parietalEEGplot.pdf", width = 5, height = 5)


eegkit::eegcap(electrodes = c("P9", "P7", "P5", "P3", "P1", "PO3", "PO7", "PO9", "POz", "Pz", "P10", "P8", "P6", "P4", "P2", "PO4", "PO8", "PO10"),  #Parietal
                              col.point = c( "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d", "#27763d"),
                             cex.point = 3, nose = TRUE, ears = TRUE,
               plotlabels = FALSE,  plotaxes = FALSE, col.head = "grey", type="2d")

dev.off()
```

```{r}

pdf("/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure2/OccipitalEEGplot.pdf", width = 5, height = 5)


eegkit::eegcap(electrodes = c("O1", "Oz", "O2","I1", "I2"),  #Occipital
                              col.point = c("#90b8a6", "#90b8a6", "#90b8a6", "#90b8a6", "#90b8a6"),
                             cex.point = 3, nose = TRUE, ears = TRUE,
               plotlabels = FALSE,  plotaxes = FALSE, col.head = "grey", type="2d")

dev.off()
```
## Rest vs Delay Boxplots

```{r}

avgACW_allEpochs <- read.csv('/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/acw/MGS/twoSecondSegments/allSubjects_allEpochs_regions.csv') %>% filter(epoch == 'Delays' | epoch == 'restEyesOpen') %>%
  select(-ACW_0)

lunaize(
  ggplot(data = avgACW_allEpochs, 
         aes(x = factor(epoch, levels = c("restEyesOpen", "Delays")), y = ACW_50, fill = interaction(epoch))) +
    geom_boxplot(alpha = 0.8, position = position_dodge(width = 0.75), color = "black", width=0.35) +
  geom_line(aes(group= interaction(lunaid)), alpha = 0.2, linewidth=0.1) +
    labs(x = "Epoch", y = "ACW", fill = "Task") +
    theme_minimal()
) + scale_fill_manual(values = c("#012b0d", "#27763d")) + 
  theme(
    legend.position = "none",
    axis.text = element_text(size = 20, color = c("black")),
    axis.title.x = element_text(size = 20, color = c("black")),
    axis.title.y = element_text(size = 20,  color = c("black")),
    axis.line = element_line(linewidth = .2), 
    axis.ticks = element_line(linewidth = .2),
    strip.background = element_rect(fill = "white")) + facet_wrap(~region)

wide <- avgACW_allEpochs %>%
  unite("epoch_region", epoch, region, sep = "_") %>%
  pivot_wider(names_from = "epoch_region", values_from = ACW_50)


t.test(wide$Delays_Frontal, wide$restEyesOpen_Frontal, paired = TRUE)
t.test(wide$Delays_Occipital, wide$restEyesOpen_Occipital, paired = TRUE)
t.test(wide$Delays_Parietal, wide$restEyesOpen_Parietal, paired = TRUE)

ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure2/INT_task_boxplot_regions.pdf", device = "pdf", dpi = 500, width = 18, height = 6)

```

## Delay vs fix
```{r}

avgACW_allEpochs <- read.csv('/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/acw/MGS/twoSecondSegments/allSubjects_allEpochs_regions.csv') %>% 
  filter(epoch == 'Delays' | epoch == 'restEyesOpen' | epoch == 'Fix') %>%
  select(-ACW_0)
```


```{r}

avgACW_allEpochs <- avgACW_allEpochs %>%
  mutate(epoch = factor(epoch, levels = c("Fix", "Delays")))  # Explicit order

lunaize(
  ggplot(
    data = avgACW_allEpochs %>% filter(epoch %in% c("Fix", "Delays")),
    aes(x = epoch, y = ACW_50, fill = epoch)
  ) +
    geom_boxplot(alpha = 0.8, position = position_dodge(width = 0.75), color = "black", width = 0.35) +
    geom_line(aes(group = interaction(lunaid)), alpha = 0.2, linewidth = 0.1) +
    labs(x = "Epoch", y = "ACW", fill = "Task") +
    theme_minimal()
) +
  scale_fill_manual(values = c("Delays" = "#012b0d", "Fix" = "#27763d")) +
  theme(
    legend.position = "right",
    axis.text = element_text(size = 20, color = "black"),
    axis.title.x = element_text(size = 20, color = "black"),
    axis.title.y = element_text(size = 20, color = "black"),
    axis.line = element_line(linewidth = 0.2),
    axis.ticks = element_line(linewidth = 0.2)
  ) +
  facet_wrap(~region)

wide <- avgACW_allEpochs %>%
  unite("epoch_region", epoch, region, sep = "_") %>%
  pivot_wider(names_from = "epoch_region", values_from = ACW_50)

# No sig differences between delay and fix in a pairwise t test 
t.test(wide$Delays_Frontal, wide$Fix_Frontal, paired = TRUE)
t.test(wide$Delays_Occipital, wide$Fix_Occipital, paired = TRUE)
t.test(wide$Delays_Parietal, wide$Fix_Parietal, paired = TRUE)

ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure4/INT_delayFix_boxplot_regions.pdf", device = "pdf", dpi = 500, width = 18, height = 6)


# across age they are basically on top of each other

lunaize(ggplot(data = avgACW_allEpochs  %>% 
  filter(epoch == 'Delays' | epoch == 'Fix'), aes(x = age, y = ACW_50, color=region, linetype=epoch))) + geom_point(aes(shape=epoch), alpha =0.3) +
  geom_line(aes(group= interaction(lunaid, epoch)), alpha = 0.2) +
  geom_smooth(aes(group=epoch), method=mgcv::"gam", formula = y ~ s(x, k = 3, fx = T), alpha=0.3, linewidth = 1.5)+ facet_wrap(~region)+
  scale_color_manual(values = c("#012b0d", "#90b8a6", "#27763d")) + 
  theme(
    legend.position = "right",
    axis.text = element_text(size = 20, color = c("black")),
    axis.title.x = element_text(size = 20, color = c("black")),
    axis.title.y = element_text(size = 20, color = c("black")),
    axis.line = element_line(linewidth = .2), 
    axis.ticks = element_line(linewidth = .2),
    strip.background = element_rect('white'))

ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure4/INT_delayFix_acrossAge_regions.pdf", device = "pdf", dpi = 500, width = 18, height = 6)


```
