==> 06_INT_developmental_effects.Rmd <==
---
title: "DevelopmentalEffects"
author: "Shane Devine McKeon"
output: html_document
date: "2025-05-30"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(LNCDR)
library(dplyr)
library(ggplot2)
library(mgcv)
library(lme4)
library(lmerTest)
library(tidyr)
library(gratia)
library(tidyverse)
library(ggeffects)
library(marginaleffects)
library(extrafont)
extrafont::loadfonts()
library(conflicted)
library(eegkit)

conflict_prefer("filter", "dplyr")  # Always use dplyr::filter()


```

## Average delay and fix to represent a overall "task state"
```{r}
avgACW_allEpochs <- read.csv('/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/acw/MGS/twoSecondSegments/allSubjects_allEpochs_regions.csv') %>% 
  filter(epoch == 'Delays' | epoch == 'restEyesOpen' | epoch == 'Fix') %>%
  select(-ACW_0)

task = avgACW_allEpochs  %>% filter(epoch == 'Delays' | epoch == 'Fix')

task_avg <- task %>%
  group_by(lunaid, region, age, sex, visitno) %>%
  summarize(
    ACW_50 = mean(ACW_50, na.rm = TRUE),
    age = mean(age, na.rm = TRUE)) %>% mutate(ageGroup = cut(age, c(0,18,Inf), labels = c('Adol','Adults')))

# task state ACW decreases across age 

lunaize(ggplot(data = task_avg, aes(x = age, y = ACW_50, color=region))) + geom_point(alpha =0.3) +
  geom_line(aes(group= interaction(lunaid)), alpha = 0.2) +
  geom_smooth(aes(group=1), method=mgcv::"gam", formula = y ~ s(x, k = 3, fx = T), alpha=0.3, linewidth = 1.5)+ facet_wrap(~region)+
  scale_color_manual(values = c("#012b0d", "#90b8a6", "#27763d")) + 
  theme(
    legend.position = "right",
    axis.text = element_text(size = 20, color = c("black")),
    axis.title.x = element_text(size = 20, color = c("black")),
    axis.title.y = element_text(size = 20, color = c("black")),
    axis.line = element_line(linewidth = .2), 
    axis.ticks = element_line(linewidth = .2),
    strip.background = element_rect('white'))

ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure4/INT_task_acrossAge_regions.pdf", device = "pdf", dpi = 500, width = 18, height = 6)


task_avg$region <- factor(task_avg$region, ordered=T)

gam.model <-  mgcv::gamm(ACW_50 ~ s(age, k = 3) + sex + region , data = task_avg, random=list(lunaid=~1))
summary(gam.model$gam)

# frontal is sig diff than occipital and parietal 
gam.model <-  mgcv::gamm(ACW_50 ~ s(age, k = 4) + s(age, k = 4, by=region) + sex + region , data = task_avg, random=list(lunaid=~1))
summary(gam.model$gam)

```

# Task and Rest across age
```{r}
lunaize(ggplot(data = task_rest, aes(x = age, y = ACW_50, color=region, linetype=epoch))) + geom_point(aes(shape=epoch), alpha =0.3) +
  geom_line(aes(group= interaction(lunaid, epoch)), alpha = 0.2) +
  geom_smooth(aes(group=epoch), method=mgcv::"gam", formula = y ~ s(x, k = 3, fx = T), alpha=0.5, linewidth = 1.5)+ facet_wrap(~region)+
  scale_color_manual(values = c("#012b0d", "#90b8a6", "#27763d")) + 
  theme(
    legend.position = "right",
    axis.text = element_text(size = 20, color = c("black")),
    axis.title.x = element_text(size = 20, color = c("black")),
    axis.title.y = element_text(size = 20, color = c("black")),
    axis.line = element_line(linewidth = .2), 
    axis.ticks = element_line(linewidth = .2),
    strip.background = element_rect('white')) + facet_wrap(~region)

ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure4/INT_taskRest_acrossAge_regions.pdf", device = "pdf", dpi = 500, width = 18, height = 6)


task_rest$region <- factor(task_rest$region, ordered=T)

gam.model <-  mgcv::gamm(ACW_50 ~ s(age, k = 3) + sex + region + epoch, data = task_rest, random=list(lunaid=~1))
summary(gam.model$gam)

# frontal is sig diff than occipital and parietal 
gam.model <-  mgcv::gamm(ACW_50 ~ s(age, k = 4) + s(age, k = 4, by=region) + sex + region + epoch , data = task_rest, random=list(lunaid=~1))
summary(gam.model$gam)


task_rest$epoch <- factor(task_rest$epoch, ordered=T)

# task is sig diff than rest 
gam.model <-  mgcv::gamm(ACW_50 ~ s(age, k = 4) + s(age, k = 4, by=epoch) + sex + epoch , data = task_rest %>% filter(region == 'Occipital'), random=list(lunaid=~1))
summary(gam.model$gam)

gam.model <-  mgcv::gamm(ACW_50 ~ s(age, k = 4) + s(age, k = 4, by=epoch) + sex + epoch , data = task_rest %>% filter(region == 'Frontal'), random=list(lunaid=~1))
summary(gam.model$gam)

gam.model <-  mgcv::gamm(ACW_50 ~ s(age, k = 4) + s(age, k = 4, by=epoch) + sex + epoch , data = task_rest %>% filter(region == 'Parietal'), random=list(lunaid=~1))
summary(gam.model$gam)
```

# Age of maturation 
```{r}

####  ACW Rate of maturation ----

epochs <- unique(task_rest$epoch)
regions <- unique(task_rest$region)
task_rest$lunaid <- as.factor(task_rest$lunaid)
task_rest$epoch <- as.factor(task_rest$epoch)

# Initialize lists to store results
gamstats_list <- list()
gamderivatives_list <- list()
gammodel_list <- list()

# Loop through each combination of epoch and region
for (task in epochs) {
  for (loc in regions) {
    
    # Apply the growthrate function for each combination
    allstats <- growthrate(task_rest %>% 
                             filter(epoch == task & region == loc),
                           'ACW_50', 'age')
    
    # Create a data frame of stats and add epoch  info
    gamstats <- allstats$gam.statistics %>% 
      mutate(epoch = task) %>%
      mutate(region = loc)
    
    # Create a data frame of derivatives and add epoch info
    gamderivatives <- allstats$gam.derivatives %>% 
      mutate(epoch = task) %>%
      mutate(region = loc)
    
    
    # Store results
    gamstats_list[[paste(task,loc, sep = "_")]] <- gamstats
    gamderivatives_list[[paste(task, loc, sep = "_")]] <- gamderivatives
    gammodel_list[[paste(task, loc, sep = "_")]] <- allstats$model
    
  }
}

# Combine results into data frames (if desired)
gamstats_combined <- bind_rows(gamstats_list, .id = "epoch_region") %>%
  separate(epoch_region, into = c("epoch", "region"), sep = "_", remove = TRUE)

gamderivatives_combined <- bind_rows(gamderivatives_list, .id = "epoch_region") %>%
  separate(epoch_region, into = c("epoch", "region"), sep = "_", remove = TRUE)

epoch.rate <- gamderivatives_combined %>% group_by(epoch, region) %>% do(rate = mean(.$derivative)) %>% unnest(cols = "rate")

epoch.rate %>% 
  ggplot(aes(y = rate, x = epoch, fill = region)) +
  geom_point(shape = 23, size = 10) +
  theme_classic() +
  xlab("\n") +
  ylab("Mean Derivative\n") +
  scale_color_manual(values = c("#012b0d", "#90b8a6", "#27763d")) +  # Custom line colors
  scale_fill_manual(values = c("#012b0d", "#90b8a6", "#27763d")) +
  theme(
    legend.position = "none",
    axis.text = element_text(size = 12,  color = c("black")),
    axis.title.x = element_text(size = 12,  color = c("black")),
    axis.title.y = element_text(size = 12,  color = c("black")),
    axis.line = element_line(linewidth = .2), 
    axis.ticks = element_line(linewidth = .2)) 

ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure3/TaskRestRateofMat.pdf", device = "pdf", dpi = 500, width = 6, height = 9)


####  ACW Age of maturation ----

epoch.age <- gamstats_combined %>% group_by(epoch, region) %>% do(age = mean(.$smooth.decrease.offset, na.rm = T)) %>% unnest(cols = "age") 

epoch.age %>% 
  ggplot(aes(x = age, y = epoch, fill = region)) +
  geom_point(shape = 23, size = 5.5, position=position_jitter(width=0.3, height=0)) +
  theme_classic() +
  xlab("\nAge of Maturation") +
  ylab("epoch\n") +
  scale_color_manual(values = c("#012b0d", "#90b8a6", "#27763d")) +  # Custom line colors
  scale_fill_manual(values = c("#012b0d", "#90b8a6", "#27763d")) +
  theme(
    legend.position = "right",
    axis.text = element_text(size = 12,  color = c("black")),
    axis.title.x = element_text(size = 12,  color = c("black")),
    axis.title.y = element_text(size = 12,  color = c("black")),
    axis.line = element_line(linewidth = .2), 
    axis.ticks = element_line(linewidth = .2)) 

ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure3/delayRestAgeofMat.pdf", device = "pdf", dpi = 500, width = 9, height = 6)


epoch.age %>% 
  ggplot(aes(y = age, ymin = 15, x = epoch, ymax = age, group = region, color = region)) +
  geom_linerange(position = position_dodge(0.5), linewidth = 3) +
  scale_y_continuous(breaks = c(16, 18, 20, 22, 24, 26, 28, 30, 32)) +
  ylab("\nAge of Maturation") +
  xlab("Epoch\n") +
  coord_flip() +
  theme_classic() +
  scale_color_manual(values = c("#012b0d", "#90b8a6", "#27763d"))+
  theme(
    legend.position = "none",
    axis.text = element_text(size = 12,  color = c("black")),
    axis.title.x = element_text(size = 12,  color = c("black")),
    axis.title.y = element_text(size = 12,  color = c("black")),
    axis.line = element_line(linewidth = .2), 
    axis.ticks = element_line(linewidth = .2)) 


ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure3/delayRestAgeofMatLines.pdf", device = "pdf", dpi = 500, width = 9, height = 6)


```

## Task vs rest boxplot
```{r}
task_avg$epoch <- as.factor("Task")
task_rest <- rbind(task_avg,  avgACW_allEpochs %>% filter(epoch == 'restEyesOpen'))
task_rest$epoch <- as.factor(task_rest$epoch)

task_rest$epochRev <- factor(task_rest$epoch, levels = rev(levels(task_rest$epoch)))
write.csv(task_rest, file = '/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/acw/taskRest.csv')

lunaize(ggplot(data = task_rest, aes(x = ageGroup, y = ACW_50, fill = interaction(factor(epoch, levels = c("restEyesOpen", "Task"))))) +
  geom_boxplot(alpha = 0.8, position = position_dodge(width = 0.75), color = "black", width = 0.35) +
  labs(x = "Epoch", y = "ACW", fill = "Task") +
  theme_minimal() + 
  scale_fill_manual(values = c("#012b0d", "#27763d", "#012b0d", "#27763d", "#012b0d", "#27763d"))) + 
  theme(
    legend.position = 'none',
    axis.text = element_text(size = 20, color = "black"),  # Removed vectorized input here
    axis.title.x = element_text(size = 20, color = "black"),  # Removed vectorized input here
    axis.title.y = element_text(size = 20, color = "black"),  # Removed vectorized input here
    axis.line = element_line(linewidth = 0.2), 
    axis.ticks = element_line(linewidth = 0.2),
    strip.background = element_rect('white')) + 
  facet_wrap(~factor(region, levels = c("Frontal", "Parietal", "Occipital")))



ggsave(filename = "/Volumes/Hera/Projects/7TBrainMech/scripts/eeg/Shane/Entropy/Results/Figures/Figure4/INT_taskRest_boxplots_regions.pdf", device = "pdf", dpi = 500, width = 18, height = 6)


wide <- task_rest %>% select(-epochRev) %>%
  unite("epoch_region", epoch, region, sep = "_") %>%
  pivot_wider(names_from = "epoch_region", values_from = ACW_50)

# sig diff between task and rest in parietal and occipital (basically the same results as with just rest)
t.test(wide$Task_Frontal, wide$restEyesOpen_Frontal, paired = TRUE)
t.test(wide$Task_Occipital, wide$restEyesOpen_Occipital, paired = TRUE)
t.test(wide$Task_Parietal, wide$restEyesOpen_Parietal, paired = TRUE)


```
